Statistics without proper frontend fix

+++ FootBookEJBProject/ejbModule/facade/Facade.java
@@ -110,11 +110,11 @@ public class Facade implements FacadeLocal {
 		userEAO.updateUser(user);
 	}
 	
-	public int getUserCount() {
+	public long getUserCount() {
 		return userEAO.getUserCount();
 	}
 	
-	public int getUsersOnMatchesCount() {
+	public long getUsersOnMatchesCount() {
 		return userEAO.getUsersOnMatchesCount();
 	}
